requirements.txt
fastapi
uvicorn
pyjwt
requests
pytest